protect routes

+++ backend/controllers/goalController.js
@@ -1,8 +1,9 @@
 const asyncHandler = require('express-async-handler')
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
 
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({ user: req.user.id })
 
     res.json(goals)
 })
@@ -14,19 +15,44 @@ const setGoal = asyncHandler(async (req, res) => {
     }
 
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
 
     res.json(goal)
 })
 
 const updateGoal = asyncHandler(async (req, res) => {
-    const goal = await Goal.findByIdAndUpdate(req.params.id, req.body)
+    const user = await User.findById(req.user.id)
 
-    res.json({ message: `Updated goal ${goal}`})
+    if (!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+    
+    if(goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
+    }
+    
+    const goal = await Goal.findByIdAndUpdate(req.params.id, req.body)
+    res.json(goal)
 })
 
 const deleteGoal = asyncHandler(async (req, res) => {
+    
+    const user = await User.findById(req.user.id)
+
+    if (!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+    
+    if(goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
+    }
+    
     const goal = await Goal.findById(req.params.id)
     
     await Goal.deleteOne(goal)
